+++ KosmicTask/KosmicTask/source/mother framework/NDScript/NSString+NDUtilities.h
@@ -124,7 +124,7 @@
 	@param range The range to limit the search to.
 	@result Returns the index of the character or <tt>NSNotFound</tt> if not found.
  */
-- (unsigned int)indexOfCharacter:(unichar)character range:(NSRange)range;
+- (NSUInteger)indexOfCharacter:(unichar)character range:(NSRange)range;
 
 
 /*!
@@ -134,7 +134,7 @@
 	@param character The character to look for.
 	@result Returns the index of the character or <tt>NSNotFound</tt> if not found.
  */
-- (unsigned int)indexOfCharacter:(unichar)character;
+- (NSUInteger)indexOfCharacter:(unichar)character;
 
 /*!
 	@method containsCharacter:
@@ -358,8 +358,8 @@ enum
 + (NSString *)stringFromDictionary:(NSDictionary *)aDictionary withFormat:(NSString *)aFormat, ...;
 + (NSString *)stringFromDictionary:(NSDictionary *)aDictionary withFormat:(NSString *)aFormat arguments:(va_list)anArguments;
 
-- (unsigned int)indexOfCharacater:(unichar)aChar;
-- (unsigned int)indexOfCharacater:(unichar)aChar options:(NSStringCompareOptions)mask;
+- (NSUInteger)indexOfCharacater:(unichar)aChar;
+- (NSUInteger)indexOfCharacater:(unichar)aChar options:(NSStringCompareOptions)mask;
 /*!
 	@method indexOfCharacater:
 	@abstract Find the index of a chartacter.
@@ -388,7 +388,7 @@ enum
 	@param range range The range to contrain the search to.
 	@result The index of the character or <tt>NSNotFound</tt> if the character</tt>
  */
-- (unsigned int)indexOfCharacater:(unichar)aChar options:(NSStringCompareOptions)mask range:(NSRange)range;
+- (NSUInteger)indexOfCharacater:(unichar)aChar options:(NSStringCompareOptions)mask range:(NSRange)range;
 
 /*!
 	@method componentsSeparatedByString:withOpeningQuote:closingQuote:singleQuote:
